chore: refactor import of Translation so it is never done via index.ts file (post-AI fixies)
