+++ packages/suite/src/components/firmware/FirmwareLowBatteryModal.tsx
@@ -1,6 +1,7 @@
 import { Card, Column, Text } from '@trezor/components';
 
-import { Translation } from '../suite';
+import { Translation } from 'src/components/suite/Translation';
+
 import { LowBatteryModal } from '../suite/modals/LowBatteryModal';
 
 type FirmwareLowBatteryModalProps = {
