+++ packages/suite/src/components/wallet/TransactionItem/TransactionHeading.tsx
@@ -14,12 +14,8 @@ import { spacings } from '@trezor/theme';
 import { HELP_CENTER_ZERO_VALUE_ATTACKS } from '@trezor/urls';
 import { BigNumber } from '@trezor/utils/src/bigNumber';
 
-import {
-    FormattedCryptoAmount,
-    TooltipSymbol,
-    Translation,
-    TrezorLink,
-} from 'src/components/suite';
+import { FormattedCryptoAmount, TooltipSymbol, TrezorLink } from 'src/components/suite';
+import { Translation } from 'src/components/suite/Translation';
 import { WalletAccountTransaction } from 'src/types/wallet';
 
 import { InstantStakeBadge } from './InstantStakeBadge';
