+++ packages/suite/src/views/dashboard/AssetsView/AssetTable/AssetStakingRow.tsx
@@ -2,12 +2,8 @@ import { NetworkSymbol } from '@suite-common/wallet-config';
 import { Column, Icon, Table, Text } from '@trezor/components';
 import { spacings } from '@trezor/theme';
 
-import {
-    BaseCurrencyValue,
-    CoinBalance,
-    HiddenPlaceholder,
-    Translation,
-} from 'src/components/suite';
+import { BaseCurrencyValue, CoinBalance, HiddenPlaceholder } from 'src/components/suite';
+import { Translation } from 'src/components/suite/Translation';
 
 import { AssetTableExtraRowsSection as Section } from './AssetTableExtraRowsSection';
