+++ packages/suite/src/views/settings/SettingsDevice/AuthenticateDevice.tsx
@@ -1,13 +1,8 @@
 import { HELP_CENTER_DEVICE_AUTHENTICATION } from '@trezor/urls';
 
 import { openModal } from 'src/actions/suite/modalActions';
-import {
-    ActionButton,
-    ActionColumn,
-    SectionItem,
-    TextColumn,
-    Translation,
-} from 'src/components/suite';
+import { ActionButton, ActionColumn, SectionItem, TextColumn } from 'src/components/suite';
+import { Translation } from 'src/components/suite/Translation';
 import { useDispatch } from 'src/hooks/suite';
 
 interface AuthenticateDeviceProps {
