+++ packages/suite/src/views/settings/SettingsDevice/BluetoothEraseBonds.tsx
@@ -2,13 +2,8 @@ import { useState } from 'react';
 
 import { selectSelectedDeviceLabelOrName } from '@suite-common/wallet-core';
 
-import {
-    ActionButton,
-    ActionColumn,
-    SectionItem,
-    TextColumn,
-    Translation,
-} from 'src/components/suite';
+import { ActionButton, ActionColumn, SectionItem, TextColumn } from 'src/components/suite';
+import { Translation } from 'src/components/suite/Translation';
 import { useDispatch, useSelector } from 'src/hooks/suite';
 
 import { bluetoothEraseBondsThunk } from '../../../actions/bluetooth/bluetoothEraseBondsThunk';
