+++ packages/suite/src/views/settings/SettingsDevice/DeviceAuthenticityOptOut.tsx
@@ -2,13 +2,8 @@ import { HELP_CENTER_DEVICE_AUTHENTICATION } from '@trezor/urls';
 
 import { openModal } from 'src/actions/suite/modalActions';
 import { toggleDeviceAuthenticityCheck } from 'src/actions/suite/suiteActions';
-import {
-    ActionButton,
-    ActionColumn,
-    SectionItem,
-    TextColumn,
-    Translation,
-} from 'src/components/suite';
+import { ActionButton, ActionColumn, SectionItem, TextColumn } from 'src/components/suite';
+import { Translation } from 'src/components/suite/Translation';
 import { useDispatch, useSelector } from 'src/hooks/suite';
 import { selectIsDeviceAuthenticityCheckEnabled } from 'src/selectors/suite/suiteSelectors';
