+++ packages/suite/src/views/settings/SettingsDevice/FirmwareTypeChange.tsx
@@ -10,13 +10,8 @@ import { HELP_FIRMWARE_TYPE } from '@trezor/urls';
 
 import { goto } from 'src/actions/suite/routerActions';
 import { SettingsSectionItem } from 'src/components/settings/SettingsSectionItem';
-import {
-    ActionButton,
-    ActionColumn,
-    TextColumn,
-    Translation,
-    TrezorLink,
-} from 'src/components/suite';
+import { ActionButton, ActionColumn, TextColumn, TrezorLink } from 'src/components/suite';
+import { Translation } from 'src/components/suite/Translation';
 import { SettingsAnchor } from 'src/constants/suite/anchors';
 import { useDevice, useDispatch } from 'src/hooks/suite';
 import { getSuiteFirmwareTypeString } from 'src/utils/firmware';
