+++ packages/suite/src/views/wallet/tokens/common/TokensTable/TokenRow.tsx
@@ -47,9 +47,9 @@ import {
     BaseCurrencyValue,
     FormattedCryptoAmount,
     PriceTicker,
-    Translation,
     TrendTicker,
 } from 'src/components/suite';
+import { Translation } from 'src/components/suite/Translation';
 import {
     useDevice,
     useDispatch,
